medihelp-ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
pytesseract==0.3.13
Pillow==10.4.0
opencv-python-headless==4.10.0.84
httpx==0.27.2
google-generativeai==0.8.2
py-eureka-client==0.11.10
